requirements.txt
boto3==1.26.92
requests==2.28.2
python-dotenv==1.0.0
pandas==1.5.3